Consolidate projection into a single stateful Project Geometry operator
Replace the dialog-based whole-object projection with the interactive
element-picking stateful operator under the existing ProjectGeometry id.
Remove the parallel ProjectElement operator, tool, enum entries and panel
button so there is one project operator, exposed as a workspace tool with
mesh-element hover feedback.
