+++ declarations.py
@@ -69,7 +69,6 @@ class Operators(str, Enum):
     NodeArrayLinear = "view3d.slvs_node_array_linear"
     NodeRevolve = "view3d.slvs_node_revolve"
     NodeBoolean = "view3d.slvs_node_boolean"
-    ProjectElement = "view3d.slvs_project_element"
     ProjectGeometry = "view3d.slvs_project_geometry"
     RegisterDrawCB = "view3d.slvs_register_draw_cb"
     RenameCurve = "view3d.slvs_rename_curve"
@@ -127,7 +126,7 @@ class WorkSpaceTools(str, Enum):
     Trim = "sketcher.slvs_trim"
     Bevel = "sketcher.slvs_bevel"
     AddSketch = "sketcher.slvs_add_sketch"
-    ProjectElement = "sketcher.slvs_project_element"
+    ProjectGeometry = "sketcher.slvs_project_geometry"
     Extrude = "sketcher.slvs_node_extrude"
     ArrayLinear = "sketcher.slvs_node_array_linear"
     Revolve = "sketcher.slvs_node_revolve"
