+++ ui/panels/tools.py
@@ -18,7 +18,6 @@ class VIEW3D_PT_sketcher_tools(VIEW3D_PT_sketcher_base):
 
         layout.operator(declarations.Operators.MergePoints)
         layout.operator(declarations.Operators.ProjectGeometry, icon="MOD_SHRINKWRAP")
-        layout.operator(declarations.Operators.ProjectElement, icon="VERTEXSEL")
 
         # Constraints
         layout.label(text="Constraints:")
