+++ workspacetools/__init__.py
@@ -10,7 +10,7 @@ from .bevel import VIEW3D_T_slvs_bevel
 from .extrude import VIEW3D_T_slvs_node_extrude
 from .manager import ToolGroup, add
 from .offset import VIEW3D_T_slvs_offset
-from .project_element import VIEW3D_T_slvs_project_element
+from .project_geometry import VIEW3D_T_slvs_project_geometry
 from .revolve import VIEW3D_T_slvs_node_revolve
 from .select import VIEW3D_T_slvs_select
 from .trim import VIEW3D_T_slvs_trim
@@ -27,7 +27,7 @@ add(VIEW3D_T_slvs_add_rectangle, visibility=ToolGroup.SKETCH, separator=False, g
 add(VIEW3D_T_slvs_trim, visibility=ToolGroup.SKETCH, separator=True, group=False)
 add(VIEW3D_T_slvs_bevel, visibility=ToolGroup.SKETCH, separator=False, group=False)
 add(VIEW3D_T_slvs_offset, visibility=ToolGroup.SKETCH, separator=False, group=False)
-add(VIEW3D_T_slvs_project_element, visibility=ToolGroup.SKETCH, separator=True, group=False)
+add(VIEW3D_T_slvs_project_geometry, visibility=ToolGroup.SKETCH, separator=True, group=False)
 
 # Non-sketch tools (visible when no sketch is active)
 add(VIEW3D_T_slvs_add_sketch, visibility=ToolGroup.NON_SKETCH, separator=True, group=False)
